+++ src/set.h
@@ -116,7 +116,7 @@ struct Set {
     }
 
     void add(T value) {
-        // Equallity is important for capacity == 0
+        // Equality is important for capacity == 0
         if (count * 10 >= capacity * GDSTK_MAP_CAPACITY_THRESHOLD)
             resize(capacity >= GDSTK_INITIAL_MAP_CAPACITY ? capacity * GDSTK_MAP_GROWTH_FACTOR
                                                           : GDSTK_INITIAL_MAP_CAPACITY);