+++ src/style.cpp
@@ -79,7 +79,7 @@ Style* StyleMap::get_slot(Tag tag) const {
 
 void StyleMap::set(Tag tag, const char* value) {
     assert(value);
-    // Equallity is important for capacity == 0
+    // Equality is important for capacity == 0
     if (count * 10 >= capacity * GDSTK_MAP_CAPACITY_THRESHOLD)
         resize(capacity >= GDSTK_INITIAL_MAP_CAPACITY ? capacity * GDSTK_MAP_GROWTH_FACTOR
                                                       : GDSTK_INITIAL_MAP_CAPACITY);