+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - `len(Library)` returns the number of cells in the library
 - `Reference.cell_name` to directly access a referenced cell's name
 
+### Fixed
+- `RobustPath` accepts width of 0 at the path end
+
 ## 0.9.37 - 2023002-12
 ### Changed
 - Build system changes for lower numpy version requirements